requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-cas-ng
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-itemlist==0.2.6
django-memoize==2.3.1
django-model-utils==4.0.0
idna==2.9
lxml==4.5.1
psycopg2-binary==2.8.5
python-cas==1.5.0
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
